feat: upgrade to the latest Cardano Client Lib, 0.6.3 which is Conway era / Ploomin HFC compatible.

+++ test-data-generator/src/main/java/org/cardanofoundation/rosetta/testgenerator/common/BaseFunctions.java
@@ -56,7 +56,7 @@ public class BaseFunctions {
     try {
       TransactionContent txContent = backendService.getTransactionService().getTransaction(txHash)
           .getValue();
-      Integer blockHeight = txContent.getBlockHeight();
+      Long blockHeight = txContent.getBlockHeight();
       log.info("Block height: {}", blockHeight);
       return backendService.getBlockService()
           .getBlockByNumber(BigInteger.valueOf(blockHeight)).getValue();
